# ELAT2026_SemanticKnowledge/local_01_vector/elat.local.llm/elat.local.llm.Server/WeatherForecast.cs
namespace elat.local.llm.Server
{
    public class WeatherForecast
    {
        public DateOnly Date { get; set; }

        public int TemperatureC { get; set; }

        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);

        public string? Summary { get; set; }
    }
}