# ELAT2026_SemanticKnowledge/local_01_vector/elat.local.llm/elat.local.llm.Server/openapi.yaml
openapi: 3.1.0
info:
  title: elat Local LLM Chat API
  description: |
    A modern chat application API built with .NET 9 that provides chat functionality with AI streaming using Ollama LLM via Semantic Kernel.
    Features include:
    - Chat sessions management
    - File/image upload support
    - Real-time AI streaming using Server-Sent Events (SSE)
    - Azure AD authentication
    - Message history persistence
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://localhost:7154
    description: Development server (HTTPS)
  - url: http://localhost:5154
    description: Development server (HTTP)

security:
  - BearerAuth: []
  - AzureAD: []

paths:
  /api/chat/sessions:
    get:
      summary: Get all chat sessions
      description: Retrieve all chat sessions for the authenticated user
      operationId: getChatSessions
      tags:
        - Chat
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSessionDto'
        '401':
          description: Unauthorized - Authentication required
        '500':
          description: Internal server error

    post:
      summary: Create a new chat session
      description: Create a new chat session for the authenticated user
      operationId: createChatSession
      tags:
        - Chat
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatSessionRequest'
      responses:
        '200':
          description: Successfully created chat session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDto'
        '400':
          description: Bad request - Invalid input
        '401':
          description: Unauthorized - Authentication required
        '500':
          description: Internal server error

  /api/chat/sessions/{sessionId}:
    get:
      summary: Get a specific chat session
      description: Retrieve a specific chat session by ID for the authenticated user
      operationId: getChatSession
      tags:
        - Chat
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: The unique identifier of the chat session
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved chat session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDto'
        '401':
          description: Unauthorized - Authentication required
        '404':
          description: Chat session not found
        '500':
          description: Internal server error

  /api/chat/sessions/{sessionId}/messages:
    post:
      summary: Send a message to a chat session
      description: |
        Send a message to a specific chat session and receive AI response via Server-Sent Events (SSE).
        This endpoint supports file uploads and returns a streaming response.
      operationId: sendMessage
      tags:
        - Chat
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: The unique identifier of the chat session
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: The message content
                  example: "Hello, how are you?"
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Optional file attachments
            encoding:
              files:
                contentType: application/octet-stream
      responses:
        '200':
          description: Message sent successfully - SSE stream response
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
              example: |
                data: {"type":"chunk","content":"Hello","messageId":"123e4567-e89b-12d3-a456-426614174000"}
                
                data: {"type":"chunk","content":" there!","messageId":"123e4567-e89b-12d3-a456-426614174000"}
                
                data: {"type":"complete","content":"","messageId":"123e4567-e89b-12d3-a456-426614174000"}
        '400':
          description: Bad request - Invalid input
        '401':
          description: Unauthorized - Authentication required
        '404':
          description: Chat session not found
        '500':
          description: Internal server error

  /api/chat/attachments/{attachmentId}:
    get:
      summary: Get an attachment
      description: Retrieve a specific attachment by ID
      operationId: getAttachment
      tags:
        - Chat
      security:
        - BearerAuth: []
      parameters:
        - name: attachmentId
          in: path
          required: true
          description: The unique identifier of the attachment
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved attachment
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized - Authentication required
        '404':
          description: Attachment not found
        '500':
          description: Internal server error

  /WeatherForecast:
    get:
      summary: Get weather forecast
      description: Get a sample weather forecast (demo endpoint)
      operationId: getWeatherForecast
      tags:
        - Weather
      responses:
        '200':
          description: Successfully retrieved weather forecast
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeatherForecast'

components:
  schemas:
    ChatSessionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the chat session
        title:
          type: string
          description: Title of the chat session
          example: "My Chat Session"
        createdAt:
          type: string
          format: date-time
          description: When the session was created
        updatedAt:
          type: string
          format: date-time
          description: When the session was last updated
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageDto'
          description: List of messages in the session
      required:
        - id
        - title
        - createdAt
        - updatedAt
        - messages

    ChatMessageDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the message
        role:
          type: string
          description: Role of the message sender
          enum: [user, assistant, system]
          example: "user"
        content:
          type: string
          description: Content of the message
          example: "Hello, how are you?"
        createdAt:
          type: string
          format: date-time
          description: When the message was created
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/ChatAttachmentDto'
          description: List of attachments for the message
      required:
        - id
        - role
        - content
        - createdAt
        - attachments

    ChatAttachmentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the attachment
        fileName:
          type: string
          description: Name of the attached file
          example: "document.pdf"
        contentType:
          type: string
          description: MIME type of the file
          example: "application/pdf"
        fileSize:
          type: integer
          format: int64
          description: Size of the file in bytes
          example: 1024
        createdAt:
          type: string
          format: date-time
          description: When the attachment was created
      required:
        - id
        - fileName
        - contentType
        - fileSize
        - createdAt

    CreateChatSessionRequest:
      type: object
      properties:
        title:
          type: string
          description: Title for the new chat session
          example: "My New Chat"
      required:
        - title

    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
          description: The message content
          example: "Hello, how are you?"
        files:
          type: array
          items:
            type: string
            format: binary
          description: Optional file attachments
      required:
        - content

    ChatStreamResponse:
      type: object
      properties:
        type:
          type: string
          description: Type of the stream response
          enum: [message, chunk, complete, error]
          example: "chunk"
        content:
          type: string
          description: Content of the response
          example: "Hello there!"
        messageId:
          type: string
          format: uuid
          description: ID of the message being streamed
        error:
          type: string
          description: Error message if type is 'error'
      required:
        - type
        - content

    WeatherForecast:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Date of the forecast
        temperatureC:
          type: integer
          description: Temperature in Celsius
          example: 25
        temperatureF:
          type: integer
          description: Temperature in Fahrenheit
          example: 77
        summary:
          type: string
          description: Weather summary
          example: "Warm"
      required:
        - date
        - temperatureC
        - temperatureF

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
      required:
        - message

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Azure AD authentication

    AzureAD:
      type: oauth2
      description: Azure Active Directory OAuth2 authentication
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            openid: OpenID Connect scope
            profile: Access to user profile information
            User.Read: Read user profile

tags:
  - name: Chat
    description: Chat session and messaging operations
  - name: Weather
    description: Weather forecast operations (demo)

externalDocs:
  description: Find more information about the API
  url: https://github.com/petkir/ESPC2025